test(Arbitrary): consolidate template literal coverage
